server/requirements.txt

alembic==1.13.3
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.52
botocore==1.35.52
click==8.1.7
coverage==7.6.4
factory-boy==3.3.0 
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
schedule==1.2.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
